requirements.txt
streamlit
python-dotenv
tqdm
requests
openai
google-genai
qdrant-client